# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Wish API",
    "description": "\nWish REST API.\n\n## Sessions\n\nSuccessful login and registration responses set an HTTP-only session cookie.\nUse it for authenticated routes such as `GET /api/auth/me`.\n\n## Email registration\n\n1. `POST /api/auth/email/send-code` — send a 6-digit code.\n2. `POST /api/auth/email/verify-register` — verify the code and create the account.\n\nWhen `SMTP_DEV_LOG_CODES=true`, the dev code is returned in the `X-Dev-Registration-Code` response header.\n",
    "version": "0.1.0"
  },
  "paths": {
    "/api/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Health check",
        "description": "Liveness probe for load balancers and monitoring.",
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register with email and password",
        "description": "Create an account and start a session (sets the `wish_session` cookie).",
        "operationId": "register_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials or missing session",
            "content": {
              "application/json": {
                "example": {
                  "detail": "invalid_credentials"
                }
              }
            }
          },
          "409": {
            "description": "Email already registered",
            "content": {
              "application/json": {
                "example": {
                  "detail": "email_taken"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/email/send-code": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Send registration verification code",
        "description": "Send a 6-digit code to the given email. Returns `Retry-After` (seconds until resend).",
        "operationId": "send_email_code_api_auth_email_send_code_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendEmailCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "Invalid, expired, or exhausted verification code",
            "content": {
              "application/json": {
                "example": {
                  "detail": "invalid_code"
                }
              }
            }
          },
          "409": {
            "description": "Email already registered",
            "content": {
              "application/json": {
                "example": {
                  "detail": "email_taken"
                }
              }
            }
          },
          "429": {
            "description": "Resend rate limit (see `Retry-After` header)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "code_rate_limited"
                }
              }
            }
          },
          "503": {
            "description": "SMTP delivery failed",
            "content": {
              "application/json": {
                "example": {
                  "detail": "email_delivery_failed"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/email/verify-register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Verify code and register",
        "description": "Verify the email code, create the account, and start a session.",
        "operationId": "verify_email_register_api_auth_email_verify_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid, expired, or exhausted verification code",
            "content": {
              "application/json": {
                "example": {
                  "detail": "invalid_code"
                }
              }
            }
          },
          "409": {
            "description": "Email already registered",
            "content": {
              "application/json": {
                "example": {
                  "detail": "email_taken"
                }
              }
            }
          },
          "429": {
            "description": "Resend rate limit (see `Retry-After` header)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "code_rate_limited"
                }
              }
            }
          },
          "503": {
            "description": "SMTP delivery failed",
            "content": {
              "application/json": {
                "example": {
                  "detail": "email_delivery_failed"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login with email and password",
        "description": "Authenticate and start a session (sets the `wish_session` cookie).",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials or missing session",
            "content": {
              "application/json": {
                "example": {
                  "detail": "invalid_credentials"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout",
        "description": "Invalidate the current session and clear the session cookie.",
        "operationId": "logout_api_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Current user",
        "description": "Return the authenticated user from the session cookie.",
        "operationId": "me_api_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials or missing session",
            "content": {
              "application/json": {
                "example": {
                  "detail": "invalid_credentials"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/config": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Auth feature flags",
        "description": "Public auth configuration for the frontend (e.g. whether Google OAuth is enabled).",
        "operationId": "auth_config_api_auth_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthConfigResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/google": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Start Google OAuth",
        "description": "Redirect the browser to Google OAuth. Callback: `GET /api/auth/google/callback`.",
        "operationId": "google_login_api_auth_google_get",
        "parameters": [
          {
            "name": "locale",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Frontend locale passed through OAuth state",
              "default": "ru",
              "title": "Locale"
            },
            "description": "Frontend locale passed through OAuth state"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AuthConfigResponse": {
        "properties": {
          "google_oauth_enabled": {
            "type": "boolean",
            "title": "Google Oauth Enabled",
            "description": "Whether Google OAuth login is configured"
          }
        },
        "type": "object",
        "required": [
          "google_oauth_enabled"
        ],
        "title": "AuthConfigResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Service status",
            "examples": [
              "ok"
            ]
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "HealthResponse"
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Account email address"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "description": "Account password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
      "RegisterRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Account email address"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password",
            "description": "At least 8 characters"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name",
            "description": "Optional display name"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "RegisterRequest"
      },
      "SendEmailCodeRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Email to send the registration code to"
          },
          "locale": {
            "type": "string",
            "pattern": "^(ru|en)$",
            "title": "Locale",
            "description": "Email language: `ru` or `en`",
            "default": "ru"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "SendEmailCodeRequest"
      },
      "UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "User UUID"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "description": "Primary email"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name",
            "description": "Display name shown in the UI"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url",
            "description": "Profile image URL"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "display_name",
          "avatar_url"
        ],
        "title": "UserResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerifyEmailCodeRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Same email used in send-code"
          },
          "code": {
            "type": "string",
            "maxLength": 6,
            "minLength": 6,
            "pattern": "^\\d{6}$",
            "title": "Code",
            "description": "6-digit verification code from email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password",
            "description": "Password for the new account"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name",
            "description": "Optional display name"
          }
        },
        "type": "object",
        "required": [
          "email",
          "code",
          "password"
        ],
        "title": "VerifyEmailCodeRequest"
      }
    }
  },
  "tags": [
    {
      "name": "health",
      "description": "Service liveness checks."
    },
    {
      "name": "auth",
      "description": "Registration, login, session cookies (`wish_session`), email verification, and Google OAuth."
    }
  ]
}